% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{clean_data}
\alias{clean_data}
\title{Removing NAs from all columns of the dataset}
\usage{
clean_data(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
dataset without NA's
}
\description{
Removing NAs from all columns of the dataset
}
\examples{
clean_data("data/ac_field_data.csv")

}
