% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafmap.R
\name{leafmap}
\alias{leafmap}
\title{Making a map of our dataset using leaflet}
\usage{
leafmap(ac_clean)
}
\arguments{
\item{ac_clean}{The dataframe you are using for your function}
}
\value{
leaflet map displaying sampling coordinates
}
\description{
Making a map of our dataset using leaflet
}
\examples{
map_data(ac_clean) 
}
