% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lin_model}
\alias{lin_model}
\title{Creating a linear model of a particular dataset}
\usage{
lin_model(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
Linear model of selected dataset
}
\description{
Creating a linear model of a particular dataset
}
\examples{
lin_model("ac_clean.csv")

}
