% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouachita_filter.R
\name{ouachita_filtered}
\alias{ouachita_filtered}
\title{Classifying the Oucahita River column}
\usage{
ouachita_filtered(ac_clean)
}
\arguments{
\item{ac_clean}{The dataframe you are using for your function}
}
\value{
fields for the ouachita river
}
\description{
Classifying the Oucahita River column
}
\examples{
clean_data(ac_clean) 
}
