% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_filter.R
\name{red_filtered}
\alias{red_filtered}
\title{Classifying the Red River column}
\usage{
red_filtered(ac_clean)
}
\arguments{
\item{ac_clean}{The dataframe you are using for your function}
}
\value{
fields for the red river
}
\description{
Classifying the Red River column
}
\examples{
red_filtered(ac_clean)
}
