% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabine_filter.R
\name{sabine_filtered}
\alias{sabine_filtered}
\title{Classifying the Sabine River column}
\usage{
sabine_filtered(ac_clean)
}
\arguments{
\item{ac_clean}{The dataframe you are using for your function}
}
\value{
fields for the sabine river
}
\description{
Classifying the Sabine River column
}
\examples{
sabine_filtered(ac_clean) 
}
