% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scatp}
\alias{scatp}
\title{Creating a scatterplot of the coordinates column of our dataset}
\usage{
scatp(ac_clean)
}
\arguments{
\item{dataframe}{}
}
\value{
Scatter plot of dataset coordinates
}
\description{
Creating a scatterplot of the coordinates column of our dataset
}
